Move build_freq_list into MerTools
